requirements-no-chromadb.txt
streamlit>=1.28.0
langchain>=0.0.350
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
openai>=1.3.0
pypdf2>=3.0.1
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.21.0
pandas>=1.3.0
tiktoken>=0.5.0
